+++ examples/base/src/widgets/index.ts
@@ -1,4 +1,0 @@
-import { Slider } from 'uiw'
-export const customWidgetsList = {
-  slider: Slider
-}