+++ packages/components/src/ProTable/BaseForm.tsx
@@ -116,6 +116,7 @@ const BaseForm: React.FC<BaseFormProps> = (props) => {
               <Col align="middle" style={{textAlign: 'right', marginRight: 14}}>
                 {searchBtns ? (
                   searchBtns.map((btn: any, idx) => (
+                    btn?.render ? <React.Fragment key={idx.toString()}>{btn.render}</React.Fragment> :
                     <Button key={idx.toString()} style={{marginRight: 5}} {...btn}>
                       {btn.label}
                     </Button>